+++ packages/shared/src/components/cards/SimilarPosts/index.ts
@@ -1,0 +1,2 @@
+export * from './PostEngagementCounts';
+export * from './SidePost';
